requirements.txt
anytree==2.12.1
app-common-python==0.2.6
bidict==0.23.0
blinker==1.7.0
cached-property==1.5.2
cachetools==5.3.2
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
crc-bonfire==5.6.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
google-auth==2.28.0
gql==3.0.0a6
graphql-core==3.1.7
greenlet==3.0.2
gunicorn==21.2.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
junitparser==3.1.1
kubernetes==29.0.0
MarkupSafe==2.1.5
multidict==6.0.5
oauthlib==3.2.2
ocviapy==1.2.3
packaging==23.2
parsedatetime==2.6
pluggy==1.4.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pytest==8.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sh==2.0.6
simple-websocket==1.0.0
six==1.16.0
tabulate==0.9.0
urllib3==2.2.0
wait-for==1.2.0
websocket-client==1.7.0
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.1
